requirements.txt
imageio
numpy
Pillow
tensorflow>=1.12.0
tqdm